添加数据展示

+++ liyun/app/src/main/java/com/android/liyun/bean/Water.java
@@ -2,15 +2,15 @@ package com.android.liyun.bean;
 
 
 public class Water {
-    private int number;
+    private double number;
     private String name;
 
-    public Water(int number, String name) {
+    public Water(double number, String name) {
         this.number = number;
         this.name = name;
     }
 
-    public int getNumber() {
+    public double getNumber() {
         return number;
     }
